Moved CommandHandler into Models
Couldn't move the dll's / ffmpeg since they need to stay in the root directory
